Update python api examples
